move generic function to sdk
